Followed a commenter's advice to make it less outrun
A reddit user on my original post noted the large amounts of overlap between "vaporwave" and "outrun", and said that it could've easily been "proper" vaporwave if there was more pastel colors. So, I interpreted it as advice and added more pastel colors and cropped the background image to hide the contrast-y checkerboard.